kernel/geom: round every product at its source (ADR-0064)
1211 products and quotients across 124 files, the torus station's
`R + r*cos v` among them. Divisions are in scope because gc
strength-reduces a divide with an exact reciprocal into a multiply BEFORE
the contraction pass, so `a + b/2` is a fusion site and `a + b/3` is not
— and gc propagates constants through variables, so no syntactic test for
"is this divisor a power of two" is sound.

Measured: `make arm64-fma PKG=./kernel/geom` 2731 -> 10. The 10 that
remain are five sites in quartic_real_roots.go, where Ferrari's factoring
is written in complex128: a complex conversion does not round (measured),
and complex division runs inside runtime.complex128div, which this repo
does not compile. Closing them means writing that solve in real
arithmetic, which moves its amd64 results; archguard's complexFusionDebt
holds the line meanwhile.

Instructions emitted: arm64 184842 -> 187413 (+1.4%), amd64 204467 ->
204461 (-0.003%). The package's 615 tests pass under arm64 emulation.

Refs #3528

+++ kernel/geom/approximate_bspline.go
@@ -60,11 +60,11 @@ func approximationKnots(ubar []float64, p, nctrl int) []float64 {
 	for i := nctrl; i <= nctrl+p; i++ {
 		knots[i] = ubar[m]
 	}
-	d := float64(m+1) / float64(n-p+1)
+	d := float64(float64(m+1) / float64(n-p+1))
 	for j := 1; j <= n-p; j++ {
 		i := int(float64(j) * d)
-		alpha := float64(j)*d - float64(i)
-		knots[p+j] = (1-alpha)*ubar[i-1] + alpha*ubar[i]
+		alpha := float64(float64(j)*d) - float64(i)
+		knots[p+j] = float64((1-alpha)*ubar[i-1]) + float64(alpha*ubar[i])
 	}
 	return knots
 }
@@ -81,10 +81,10 @@ func normalSystem(pts [][]float64, ubar []float64, p, nctrl int, knots []float64
 		rk := residual(pts[k], pts[0], pts[m], nk[0], nk[n])
 		for i := 1; i <= n-1; i++ {
 			for c := range dim {
-				rhs[i-1][c] += nk[i] * rk[c]
+				rhs[i-1][c] += float64(nk[i] * rk[c])
 			}
 			for j := 1; j <= n-1; j++ {
-				ntn[i-1][j-1] += nk[i] * nk[j]
+				ntn[i-1][j-1] += float64(nk[i] * nk[j])
 			}
 		}
 	}
@@ -108,7 +108,7 @@ func denseBasis(u float64, p, nctrl int, knots []float64) []float64 {
 func residual(qk, q0, qm []float64, n0, nn float64) []float64 {
 	r := make([]float64, len(qk))
 	for c := range qk {
-		r[c] = qk[c] - n0*q0[c] - nn*qm[c]
+		r[c] = qk[c] - float64(n0*q0[c]) - float64(nn*qm[c])
 	}
 	return r
 }
