+++ kernel/geom/bspline_curve.go
@@ -44,7 +44,7 @@ func (c BSplineCurve) PointAt(t float64) math.Point3 {
 	var h homog
 	for k := 0; k <= c.Degree; k++ {
 		i := span - c.Degree + k
-		h.add(c.Ctrl[i], c.Weights[i]*n[k])
+		h.add(c.Ctrl[i], float64(c.Weights[i]*n[k]))
 	}
 	return h.point()
 }
@@ -56,8 +56,8 @@ func (c BSplineCurve) TangentAt(t float64) math.Vector3 {
 	var value, deriv homog
 	for k := 0; k <= c.Degree; k++ {
 		i := span - c.Degree + k
-		value.add(c.Ctrl[i], c.Weights[i]*n[k])
-		deriv.add(c.Ctrl[i], c.Weights[i]*dn[k])
+		value.add(c.Ctrl[i], float64(c.Weights[i]*n[k]))
+		deriv.add(c.Ctrl[i], float64(c.Weights[i]*dn[k]))
 	}
 	return value.deriv(deriv)
 }
