+++ kernel/geom/bspline_remove.go
@@ -144,14 +144,14 @@ func rationalRemovalTol(tol float64, maxNorm float64, weights []float64) float64
 			minW = w
 		}
 	}
-	return tol * minW / (1 + maxNorm)
+	return float64(tol * minW / (1 + maxNorm))
 }
 
 // ctrlNorms3 is the largest control-point distance from the origin.
 func ctrlNorms3(ctrl []math.Point3) float64 {
 	m := 0.0
 	for _, p := range ctrl {
-		if d := stdmath.Sqrt(float64(p.X*p.X + p.Y*p.Y + p.Z*p.Z)); d > m {
+		if d := stdmath.Sqrt(float64(float64(p.X*p.X) + float64(p.Y*p.Y) + float64(p.Z*p.Z))); d > m {
 			m = d
 		}
 	}
@@ -162,7 +162,7 @@ func ctrlNorms3(ctrl []math.Point3) float64 {
 func ctrlNorms2(ctrl []math.Point2) float64 {
 	m := 0.0
 	for _, p := range ctrl {
-		if d := stdmath.Sqrt(float64(p.X*p.X + p.Y*p.Y)); d > m {
+		if d := stdmath.Sqrt(float64(float64(p.X*p.X) + float64(p.Y*p.Y))); d > m {
 			m = d
 		}
 	}
