+++ kernel/geom/canal_arc.go
@@ -59,13 +59,13 @@ func arcControls(m, fa, fb math.Point3, radius float64) (shoulder math.Point3, w
 			"crossSectionArc: foot coincides with center m=%v (|m→fa|=%g, |m→fb|=%g), expected both = radius %g",
 			m, la, lb, radius)
 	}
-	half := 0.5 * stdmath.Acos(math.Clamp(float64(da.Dot(db))/(la*lb), -1, 1))
+	half := float64(0.5 * stdmath.Acos(math.Clamp(float64(float64(da.Dot(db))/(la*lb)), -1, 1)))
 	if half <= arcMinHalfAngle || half >= halfPi-arcMinHalfAngle {
 		return math.Point3{}, 0, fmt.Errorf(
 			"crossSectionArc: feet+center near-collinear (arc half-angle %g rad outside (%g, %g)): "+
 				"grazing/antipodal, no radius-%g arc plane", half, arcMinHalfAngle, halfPi-arcMinHalfAngle, radius)
 	}
 	weight = stdmath.Cos(half)
-	bis := da.Scale(1 / la).Add(db.Scale(1 / lb)) // feet bisector → the minor-arc (cavity) side, |bis| = 2·weight
-	return m.TranslateBy(bis.Scale(radius / (2 * weight * weight))), weight, nil
+	bis := da.Scale(float64(1 / la)).Add(db.Scale(float64(1 / lb))) // feet bisector → the minor-arc (cavity) side, |bis| = 2·weight
+	return m.TranslateBy(bis.Scale(float64(radius / (2 * weight * weight)))), weight, nil
 }
